requirements.txt
discord.py[voice]
pip
youtube_dl
ffmpeg
asyncio
colorlog